renamed classes pull request
