+++ GitHubRepo/app/src/main/java/net/paulacr/githubrepo/pullrequests/PullRequestsActivity.java
@@ -1,4 +1,4 @@
-package net.paulacr.githubrepo.showpullrequests;
+package net.paulacr.githubrepo.pullrequests;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -7,7 +7,7 @@ import android.support.v7.app.AppCompatActivity;
 /**
  * Created by paularosa on 3/31/16.
  */
-public class ShowPullRequestsActivity extends AppCompatActivity{
+public class PullRequestsActivity extends AppCompatActivity{
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
